+++ NEWS.md
@@ -1,6 +1,7 @@
 # teal 0.9.0.9000
 
 * Added new function `csv_dataset_connector` and `csv_cdisc_dataset_connector`.
+* Updated `set_ui_input` method of `RawDatasetConnector` and `NamedDatasetConnector` to handle user defined shiny inputs.
 * Include `Keep Inf` checkbox for numerical filter items. `Keep NA` and `Keep Inf` checkbox doesn't 
 appear if there are no missing or infinite values.
 * Replace existing `RelationalData` class with abstract class `RelationalDataCollection` and